+++ apps/docs/components/pager.tsx
@@ -2,7 +2,7 @@ import { docsConfig } from "@/config/docs";
 import { cn } from "@/lib/utils";
 import { buttonVariants } from "@/registry/default/ui/button";
 import { ChevronLeftIcon, ChevronRightIcon } from "@radix-ui/react-icons";
-import { Doc } from "contentlayer/generated";
+import { Doc } from "contentlayer2/generated";
 import Link from "next/link";
 import { NavItem, NavItemWithChildren } from "types/nav";
 
@@ -44,7 +44,7 @@ export function DocsPager({ doc }: DocsPagerProps) {
 export function getPagerForDoc(doc: Doc) {
   const flattenedLinks = [null, ...flatten(docsConfig.sidebarNav), null];
   const activeIndex = flattenedLinks.findIndex(
-    (link) => doc.slug === link?.href
+    (link) => doc.slug === link?.href,
   );
   const prev = activeIndex !== 0 ? flattenedLinks[activeIndex - 1] : null;
   const next =
@@ -53,7 +53,7 @@ export function getPagerForDoc(doc: Doc) {
       : null;
   return {
     prev,
-    next
+    next,
   };
 }