+++ apps/docs/next.config.mjs
@@ -1,4 +1,4 @@
-import { createContentlayerPlugin } from "next-contentlayer";
+import { withContentlayer } from "next-contentlayer2";
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
@@ -10,79 +10,79 @@ const nextConfig = {
       { protocol: "https", hostname: "images.unsplash.com" },
       {
         protocol: "https",
-        hostname: "sikka-images.s3.ap-southeast-1.amazonaws.com"
-      }
-    ]
+        hostname: "sikka-images.s3.ap-southeast-1.amazonaws.com",
+      },
+    ],
   },
   redirects() {
     return [
       {
         source: "/elements",
         destination: "/docs/elements/accordion",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/docs/elements",
         destination: "/docs/elements/accordion",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/blocks",
         destination: "/docs/blocks/login-form",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/docs/blocks",
         destination: "/docs/blocks/login-form",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/layout",
         destination: "/docs/layout/app-layout",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/docs/layout",
         destination: "/docs/layout/app-layout",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/hooks",
         destination: "/docs/hooks/use-toast",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/docs/hooks",
         destination: "/docs/hooks/use-toast",
-        permanent: true
+        permanent: true,
       },
       {
         source: "/examples",
         destination: "/examples/mail",
-        permanent: false
+        permanent: false,
       },
       {
         source: "/docs/primitives/:path*",
         destination: "/docs/elements/:path*",
-        permanent: true
+        permanent: true,
       },
 
       {
         source: "/docs/forms",
         destination: "/docs/elements/form",
-        permanent: false
+        permanent: false,
       },
       {
         source: "/docs/forms/react-hook-form",
         destination: "/docs/elements/form",
-        permanent: false
-      }
+        permanent: false,
+      },
     ];
-  }
+  },
 };
 
-const withContentlayer = createContentlayerPlugin({
-  // Additional Contentlayer config options
-});
+// const withContentlayer = createContentlayerPlugin({
+//   // Additional Contentlayer config options
+// });
 
 export default withContentlayer(nextConfig);