+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hawa-storybook
 
+## 0.26.91
+
+### Patch Changes
+
+- Updated dependencies
+  - @sikka/hawa@0.36.0
+
 ## 0.26.90
 
 ### Patch Changes